Wrap each handler when added to the router
When testing a m.Use("/", logger.New()), then logger was printing output
as soon as the first middleware anywhere on the route had finished,
therefore the middlewares weren't being done properly. This now wraps
any route with all of it's middleware when being added to the router
and not during routing. Using the logger now works correctly.